requirements.txt
# requirements.txt

streamlit==1.36.0
openai==1.16.2
langchain==0.1.14
python-dotenv==1.0.1
reportlab==4.0.8
pypdf==4.2.0
requests==2.31.0
tiktoken==0.7.0
